feat: some last tweaks
